% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{anaCache}
\alias{anaCache}
\title{Analysis Cache}
\description{
Internal SetFisher field holding cached analysis objects
}
\details{
A list structure that holds previously requested SetFisherAnalysis
objects to allow more rapid recovery and reuse if they are
requested multiple times.

This field is not intended to be directly used. Attempts to
manipulate it will likely lead to unpredictable and undesired
behavior. It is built internally through calls to
\link{analysis}.
}
\seealso{
\link{analysis}, \link{SetFisherAnalysis}
}
