% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{analysesFromDirectory}
\alias{analysesFromDirectory}
\title{Get Analyses from Directory}
\arguments{
\item{dir}{Required, the path to the directory to process}

\item{query}{Default NULL, passed on to \link{analysis}}

\item{idmap}{Default NULL, passed on to \link{analysis}}

\item{pattern}{Default NULL, which will process all files in
\code{dir}. Otherwise will be passed to \link[base]{list.files}}

\item{skip}{Default NULL. Can be an optional regular expresion
which when matched against a file name will cause that file to
be skipped.}

\item{\dots}{Passed to \link{analysis}}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A list, with each member being a \code{SetFisherAnalysis}
    object.
}
\description{
Build analysis objects from all ontology matrices in a directory
}
\details{
\preformatted{
## Method Usage:
myObject$analysesFromDirectory( help=TRUE )

anas <- myObject$analysesFromDirectory(dir, query=NULL, idmap=NULL,
                                       pattern=NULL, skip=NULL, ...)
}

A convienence method that will take pattern-defined matrices from a
directory and make an analysis object for each one.
}
\seealso{
\link{analysis}
}
