% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{analysis}
\alias{analysis}
\title{Get SetFisher Analysis}
\arguments{
\item{ontology}{Required, an identifier specifying an Ontology (a
matrix defining one or more gene sets). The identifier can be a
path to the AnnotatedMatrix file, or the name of the ontology
(if it has already been loaded). If ontology is an integer, it
is instead interpreted to be an index value for the internal
analysis cache.}

\item{query}{Default \code{NULL}, in which case the value defined
by \code{defaultQuery} (if any) will be used. Otherwise and
AnnotatedMatrix (as recovered by \link{annotatedMatrix}) must
be provided.}

\item{idmap}{Default \code{NULL}, an optional ID mapping
matrix. When NULL, the matrix (if any) stored by
\link{defaultMap} will be used.}

\item{name}{Default \code{NULL}, an optional human-readable name to
assign to the analysis. If not provided, then a name will be
generated from the name assigned to the ontology.}

\item{queryworld}{Default \code{NULL}, an optional vector of ID
names to help define (restrict or expand) the "world" that the
query IDs come from. If not provided, will be taken from
\link{defaultQueryWorld}.}

\item{\dots}{Passed to \link{SetFisherAnalysis}}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A \link{SetFisherAnalysis} object.
}
\description{
Get or create a SetFisher Analysis object
}
\details{
\preformatted{
## Method Usage:
myObject$analysis( help=TRUE )

ana <- myObject$analysis(ontology, query=NULL, idmap=NULL,
                         name=NULL, queryworld=NULL, ... ) 
}

This function is the primary method within SetFisher. It creates a
new analysis object, or recovers one that was already created.

Note that this function does not actually do any of the analysis,
it just mediats the combination of pieces needed to make one (A
query matrix, an ontology matrix, and optionally a mapping matrix)
and returns an object that will itself be used to analyze for set
enrichment.
}
\examples{

\dontrun{
sf  <- SetFisher()
myMat <- sf$annotatedMatix("/data/matrices/Chickens_vs_Cows.mtx")
}

}
\seealso{
\link{SetFisherAnalysis}
}
