% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{analysisNames}
\alias{analysisNames}
\title{Compact Analysis Names}
\arguments{
\item{trim.names}{Default \code{TRUE}, which will cause the "left"
and "right" sides of all the analysis names to be trimmed to
remove common words. If FALSE, the method isn't too useful.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A character vector of analysis names.
}
\description{
Takes a set of analysis and shortens their names by removing 'common bits'
}
\details{
\preformatted{
## Method Usage:
myObject$analysisNames( help=TRUE )

anaNames <- myObject$analysisNames( trim.names=TRUE, ... )
}

While this might appear somewhat frivolous, it's intended to help
compact reports that are already rather bulky. If trimming is
requested (the default), it's designed to turn name lists like
this:

\itemize{
    \item Human Entrez GeneOntology Analysis
    \item Human Entrez MSigDB - Oncogenic Signatures Analysis
    \item Human Entrez WikiPathways Analysis
}

to this:

\itemize{
    \item GeneOntology
    \item MSigDB - Oncogenic Signatures
    \item WikiPathways
}
}
\seealso{
\link{dotSharedSubstring}
}
