% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{annotatedMatrix}
\alias{annotatedMatrix}
\title{Get AnnotatedMatrix}
\arguments{
\item{x}{Default \code{NULL}, which will simply return the default
query world, or NULL if none is set. Otherwise, will expect a
character vector of IDs that defines the world of query IDs you
wish to use.}

\item{\dots}{Passed to \link{AnnotatedMatrix}}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
\code{NULL} if none set, or a character vector of IDs that
    defines the world being analyzed.
}
\description{
Get an AnnotatedMatrix, either from a cache or de novo
}
\details{
\preformatted{
## Method Usage:
myObject$annotatedMatrix( help=TRUE )

myMat <- myObject$annotatedMatrix( x=NA, ... )
}

This is a light-weight wrapper around an \code{AnnotatedMatrix()}
call, with some additional functionality that an already-loaded
matrix will be recovered from an internal cache rather than
reloaded.
}
\examples{

\dontrun{
sf  <- SetFisher()
myMat <- sf$annotatedMatix("/data/matrices/Chickens_vs_Cows.mtx")
}

}
\seealso{
\link{AnnotatedMatrix}, \link{matrixCache}
}
