% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{datatable}
\alias{datatable}
\title{Data Table}
\arguments{
\item{res}{Default \code{NA}, which will use the results in
\link{lastTop}, if available, or call \link{processAll} if
not. Otherwise the user may provide a specific data.frame to
process.}

\item{caption}{Default \code{NULL}, which will attempt to make a
meaningful caption for the table.}

\item{\dots}{Passed to \link{processAll} if \code{res} was not
provided, and also passed to \link[dynamictable]{dynamictable}.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
The data.table object
}
\description{
Create a data.table from a top results data.frame
}
\details{
\preformatted{
## Method Usage:
myObject$datatable( help=TRUE )

dt <- myObject$datatable(res=NA, caption=NULL, ... )
}

Generate a data.table from a data.frame generated by
\link{topResults}, decorated with metadata taken from query and
ontology matrices. This uses the
\link[SetFisherAnalysis]{datatable} method built into
SetFisherAnalysis objects.
}
\seealso{
\link{topResults}, \link[SetFisherAnalysis]{datatable}
}
