% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{defaultMap}
\alias{defaultMap}
\title{Default Map Matrix}
\arguments{
\item{x}{Default \code{NULL}, which will simply return the default
map matrix, or NULL if none is set. Otherwise, will attempt
to recover the AnnotatedMatrix corresponding to \code{x} and
set it as the default.}

\item{\dots}{Passed to \link{annotatedMatrix}}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
\code{NULL} if none set, or an \link{AnnotatedMatrix}
}
\description{
Set/Get an optional default Map Matrix
}
\details{
\preformatted{
## Method Usage:
myObject$defaultMap( help=TRUE )

mapMat <- myObject$defaultMap( x=NULL, ... )
}

If a default mapping matrix is set, it will be automatically used
in cases where one is needed but none is supplied (generally when
creating a \link{SetFisherAnalysis})

The map is an optional matrix that converts IDs from one
"namespace" to another. It is used to connect queries to ontologies
when they do not share a namespace. For example, if your queries
are sets of Affymetrix probe sets, but your ontology is defined
using Ensembl gene accessions, you will need a map that associates
Affy probe set IDs with the appropriate Ensembl Gene IDs.
}
\examples{

sf  <- SetFisher()
s2e <- AnnotatedMatrix( annotatedMatrixExampleFile() )
sf$defaultMap( s2e )

}
\seealso{
\link{defMapObj}, \link{defaultQuery},
    \link{annotatedMatrix}, \link{AnnotatedMatrix}
}
