% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{defaultQuery}
\alias{defaultQuery}
\title{Default Query Matrix}
\arguments{
\item{x}{Default \code{NULL}, which will simply return the default
query matrix, or NULL if none is set. Otherwise, will attempt
to recover the AnnotatedMatrix corresponding to \code{x} and
set it as the default.}

\item{\dots}{Passed to \link{annotatedMatrix}}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
\code{NULL} if none set, or an \link{AnnotatedMatrix}
}
\description{
Set/Get an optional default Query Matrix
}
\details{
\preformatted{
## Method Usage:
myObject$defaultQuery( help=TRUE )

queryMat <- myObject$defaultQuery( x=NULL, ... )
}

If a default query matrix is set, it will be automatically used in
cases where one is needed but none is supplied (generally when
creating a \link{SetFisherAnalysis})

The query is a set of one or more sets of IDs that will be tested
for enrichment in one or more ontologies.
}
\examples{

sf  <- SetFisher()
s2e <- AnnotatedMatrix( annotatedMatrixExampleFile() )
sf$defaultQuery( s2e )

}
\seealso{
\link{defQryObj}, \link{defaultQueryWorld}, \link{defaultMap},
    \link{annotatedMatrix}, \link{AnnotatedMatrix}
}
