% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{defaultQueryWorld}
\alias{defaultQueryWorld}
\title{Default Query World}
\arguments{
\item{x}{Default \code{NULL}, which will simply return the default
query world, or NULL if none is set. Otherwise, will expect a
character vector of IDs that defines the world of query IDs you
wish to use.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
\code{NA} if none set, or a character vector of IDs that
    defines the world being analyzed.
}
\description{
Set/Get an optional default Query World
}
\details{
\preformatted{
## Method Usage:
myObject$defaultQueryWorld( help=TRUE )

myWorld <- myObject$defaultQueryWorld( x=NULL )
}

If a default query world is set, it will be automatically used
in cases where one is needed but none is supplied (generally when
creating a \link{SetFisherAnalysis})

The query world is used to constrict or expand the collection of
\strong{all possible} query IDs. For example, you may be running an
analysis on mouse genes, but your experimental proceedure
explicitly sampled a smaller subset of specific genes. In this
case, it is important to constrain the world to that set of genes
to properly calculate enrichment statistics.
}
\examples{

sf  <- SetFisher()
world <- c("ABC", "XYZ", "ZF12", "HOX4", "OCT4")
sf$defaultQueryWorld( world )
}
\seealso{
\link{defQryWorld}, \link{defaultQuery}
}
