% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{defMapObj}
\alias{defMapObj}
\title{Default Map Object}
\description{
Internal SetFisher field holding the Default Map Matrix
}
\details{
Either NULL (technically an empty field object) if no default has
been defined, or an AnnotatedMatrix object. Set using
\link{defaultMap}.

This field is not intended to be directly used. Attempts to
manipulate it will likely lead to unpredictable and undesired
behavior. Please access it using the \link{defaultMap} method.
}
\examples{

sf  <- SetFisher()
s2e <- AnnotatedMatrix( annotatedMatrixExampleFile() )
sf$defaultMap( s2e )
str( sf$defMapObj )
}
\seealso{
\link{defaultMap}
}
