% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{defQryObj}
\alias{defQryObj}
\title{Default Query Object}
\description{
Internal SetFisher field holding the Default Query Matrix
}
\details{
Either NULL (technically an empty field object) if no default has
been defined, or an AnnotatedMatrix object. Set using
\link{defaultQuery}.

This field is not intended to be directly used. Attempts to
manipulate it will likely lead to unpredictable and undesired
behavior. Please access it using the \link{defaultQuery}
method.
}
\examples{

sf  <- SetFisher()
s2e <- AnnotatedMatrix( annotatedMatrixExampleFile() )
sf$defaultQuery( s2e )
str( sf$defQryObj )
}
\seealso{
\link{defaultQuery}
}
