% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{defQueryWorld}
\alias{defQueryWorld}
\title{Default Query World}
\description{
Internal SetFisher field holding the default set of query IDs
}
\details{
Either NULL (technically an empty field object) if no default has
been defined, or a character vector of Query IDs. Set with
\link{defaultQueryWorld}.

This field is not intended to be directly used. Attempts to
manipulate it will likely lead to unpredictable and undesired
behavior. Please access it using the \link{defaultQueryWorld}
method.
}
\examples{

sf  <- SetFisher()
world <- c("ABC", "XYZ", "ZF12", "HOX4", "OCT4")
sf$defaultQueryWorld( world )
str( sf$defQueryWorld )
}
\seealso{
\link{defaultQueryWorld}, \link{defaultQuery}
}
