% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{discardedIDs}
\alias{discardedIDs}
\title{Discarded IDs List}
\description{
Internal SetFisherAnalysis field holding query, mapping and ontology IDs that are discarded / ignored / suppressed
}
\details{
This field notes any identifiers from the Query, Mapping or
Ontology matrices that are no longer "available" in the
analysis. It is populated by the \link{.noteDiscarded} method,
which is called exclusively by \link{.updateDerivedStructures}

Identifiers that were previously filtered out using AnnotatedMatrix
filter methods will not be included here.

TODO - Extend documentation!!!
}
