% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{dotNoteDiscarded}
\alias{dotNoteDiscarded}
\alias{.noteDiscarded}
\title{Record Discarded IDs}
\description{
Internal method that notes IDs discarded during structure processing
}
\details{
This method will return the \link{queryOnto} matrix if needed, and
then return it.

TODO - Extend documentation!!!
}
\keyword{internal}
