% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetFisher.R
\name{dotSharedSubstring}
\alias{dotSharedSubstring}
\alias{.sharedSubstring}
\title{Shared Substring}
\usage{
.sharedSubstring(x, left.side = TRUE)
}
\arguments{
\item{x}{Required, a character vector of strings}

\item{left.side}{Default \code{TRUE}, which will process the left
side of the strings. If FALSE it will ... wait for it
... process the \strong{RIGHT} side of the string.}
}
\value{
A string representing the common part of all strings.
}
\description{
Find a common substring on the left or right of a set of strings
}
\details{
Will identify characters common to all provided strings on either
the left or right side. Used for trimming away names to produce
more compact reports.
}
\keyword{internal}
