% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{dotUpdateDerivedStructures}
\alias{dotUpdateDerivedStructures}
\alias{.updateDerivedStructures}
\title{Update Derived Structures}
\description{
Internal method that generates derivative matrices and vectors
}
\details{
The two or three matrices used by a SetFisherAnalysis object are
pre-processed into a set of derived structures that will be used to
perform hypergeometric calculations. This method will create these
structures if they do not yet exist, or if changes have been made
to the primary matrices (generally by filtering).

TODO - Extend documentation!!!
}
\keyword{internal}
