% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetFisherUtils.R
\name{filesToMTX}
\alias{filesToMTX}
\title{Files to MatrixMarket}
\usage{
filesToMTX(files, output = "SparseMatrix.mtx", filter = NULL,
  rename = NULL)
}
\arguments{
\item{files}{Required, a vector of file paths}

\item{output}{Default "SparseMatrix.mtx", the filename of the
MatrixMarket file being generated}

\item{filter}{Optional function used to filter set elements. It
should take as input a character vector and return a character
vector representing the desired (kept) elements.}

\item{rename}{Optional function used to generate list names. It
should take as input the file path (single string) and return a
single string that will be used as a name for that list}
}
\description{
Converts a collection of files to a single MatrixMarket file
}
\examples{

\dontrun{
  # Find all files names "geneList#.txt" and extact Ensembl Gene IDs
  # from them into a single MTX file
  filesToMTX(files=list.files(pattern='geneList[0-9]+.txt'),
             filter=function(x) x[ grepl("^ENSG\\\\d+$", x) ],
             output="myEnsemblLists.mtx")
}

}

