% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{htmltable}
\alias{htmltable}
\title{HTML Table}
\arguments{
\item{res}{Default \code{NA}, which will use the results in
\link{lastTop}, if available, or call \link{processAll} if
not. Otherwise the user may provide a specific data.frame to
process.}

\item{file}{Default \code{NA}, which will generate a temporary
file. Otherwise, a specified file path may be provided.}

\item{caption}{Default \code{NULL}, which will attempt to make a
meaningful caption for the table.}

\item{show.attr}{Default \code{FALSE}, passed on to
\link[dynamictable]{dynamictable}. The default prevents
data.frame attributes from being printed on the page.}

\item{cols}{The columns to include in the table.}

\item{ontology.stats}{Default \code{TRUE}. Mysterious unimplemented
feature. Perhaps I'll remember what I was planning while
driving home.}

\item{footer}{Default an empty charcter vector. A custom table
footer can be provided if desired. Passed on to
\link[dynamictable]{dynamictable}.}

\item{style}{Default an empty charcter vector. Custom CSS styles
can be provided if desired. Passed on to
\link[dynamictable]{dynamictable}.}

\item{force}{Default \code{FALSE}. Passed on to \link{processAll}.}

\item{favicon}{An explicit \code{data:image/png;base64}
representation of a 16x16 icon to show in the browser tab when
the table is rendered.}

\item{\dots}{Passed to \link{processAll} if \code{res} was not
provided, and also passed to \link[dynamictable]{dynamictable}.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
The path to the HTML file generated.
}
\description{
Create a dynamic HTML table from a top results data.frame
}
\details{
\preformatted{
## Method Usage:
myObject$htmltable( help=TRUE )

htmlPath <- myObject$htmltable(res=NA, file=NA, caption=NULL, show.attr=FALSE,
   cols=c("term","term Description", "listName", "logPV", "RawLogPV",
          "i","N","i/N","n","W","n/W","Enrich"), ontology.stats=TRUE,
          footer=character(), style=character(), force=FALSE,
          favicon = "data:Chunk o' Base64 here", ... )
}

Create a dynamic HTML table from a data.frame generated by
\link{topResults}, decorated with metadata taken from query and
ontology matrices.
}
\seealso{
\link{topResults}
}
