% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetFisherUtils.R
\name{is.def}
\alias{is.def}
\title{Is Defined}
\usage{
is.def(x)
}
\arguments{
\item{x}{The object to be tested}
}
\value{
FALSE for NULL, objects with only NAs, 0x0 matrices and
    fields that are classed as 'uninitializedField'. All other
    values will be TRUE.
}
\description{
Broad test returning false for a variety of "empty" values
}

