% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{lastTop}
\alias{lastTop}
\title{Last Top Results}
\description{
Internal SetFisher field holding the most recent resutls from $topResults
}
\details{
A data.frame holding the most recent results from
\link{topResults}. Designed to allow "efficient" storage of the
results (since this package is a ReferenceClass object) for rapid
recovery as needed.
}
\seealso{
\link{topResults}
}
