% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{mapWeights}
\alias{mapWeights}
\title{Map Weights Matrix}
\description{
Internal SetFisherAnalysis field, a fractional map from query to ontology IDs
}
\details{
This field generally should not be used directly. Instead, using
the \link{weightMatrix} method will return this value after
checking if it needs to be recalculated.

TODO - Extend documentation!!!
}
