% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{matrixCache}
\alias{matrixCache}
\title{Matrix Cache}
\description{
Internal SetFisher field holding cached matrices
}
\details{
A list structure that holds previously requested matrices to allow
more rapid recovery and reuse if they are requested multiple times.

This field is not intended to be directly used. Attempts to
manipulate it will likely lead to unpredictable and undesired
behavior. It is built internally through calls to
\link{annotatedMatrix}.
}
\seealso{
\link{annotatedMatrix}, \link{AnnotatedMatrix}
}
