% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{processAll}
\alias{processAll}
\title{Process All Analyses}
\arguments{
\item{\dots}{Passed to \link{topResults}}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
The output from \link{topResults}
}
\description{
Filter, process, adjust and report all analyses held by SetFisher
}
\details{
\preformatted{
## Method Usage:
myObject$processAll( help=TRUE )

topRes <- myObject$processAll( ... )
}

A convienence method that will process all the SetFisherAnalysis
objects held by SetFisher, and return a topResults report for them.
}
\seealso{
\link{topResults}, \link[SetFisherAnalysis]{filter},
    \link[SetFisherAnalysis]{processListOfLists},
    \link[SetFisherAnalysis]{adjustResults}
}
