% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{queryOnto}
\alias{queryOnto}
\title{Query to Ontology Matrix}
\description{
Internal SetFisherAnalysis field, a direct query-to-ontology weighted matrix
}
\details{
This field generally should not be used directly. Instead, using
the \link{} method will return this value after
checking if it needs to be recalculated.

TODO - Extend documentation!!!
}
