% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{queryToOntology}
\alias{queryToOntology}
\title{Generate Query To Ontology}
\description{
Generate weighted query-to-ontology matrix if needed, then return it
}
\details{
This method will return the \link{queryOnto} matrix if needed, and
then return it.

TODO - Extend documentation!!!
}
