% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetFisher.R
\docType{class}
\name{SetFisher-class}
\alias{SetFisher}
\alias{SetFisher-class}
\title{SetFisher}
\description{
Enrichment analysis using hypergeometric distribution and adjusting
counts for multiple voting
}
\section{Fields}{

\describe{
\item{\code{log}}{The SetFisherLogger object holding log (activity) entries}

\item{\code{defQryObj}}{Optional default Query object (SetFisherMatrix)}

\item{\code{defMapObj}}{Optional default Map object (SetFisherMatrix)}

\item{\code{defWorld}}{Optional default World (character vector)}

\item{\code{lastTop}}{Last results from topResults()}

\item{\code{matrixCache}}{Internal storage for SetFisherMatrix objects}

\item{\code{anaCache}}{Internal storage for SetFisherAnalysis objects}
}}

