% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetFisherAnalysis.R
\docType{class}
\name{SetFisherAnalysis-class}
\alias{SetFisherAnalysis}
\alias{SetFisherAnalysis-class}
\title{SetFisher Analysis}
\description{
Class object associating the components of an enrichment analysis
(query lists, ontology, and optional mapping matrix) with the
results
}
\section{Fields}{

\describe{
\item{\code{setfisher}}{Pointer back to the parent SetFisher object}

\item{\code{log}}{The SetFisherLogger object holding log (activity)
entries}

\item{\code{queryObj}}{SetFisherMatrix representing the query list(s)}

\item{\code{queryWorld}}{Character verctor describing all objects in the
query namespace. Will default to all unique rows in the mapping
matrix, or all unique rows in the query matrix if no mapping
matrix is present.}

\item{\code{ontoObj}}{SetFisherMatrix representing the mapping of IDs
(from the mapping matrix if present, otherwise from the query
matrix) to ontology terms}

\item{\code{mapObj}}{Optional SetFisherMatrix mapping IDs in the Query
matrix to those used by the Ontology matrix}

\item{\code{mapWeights}}{Weight matrix used to accomodate multiple voting
from the Query namespace to the Ontology namespace}

\item{\code{queryUse}}{Boolean matrix associating queries with
lists. Intended to allow threshold filters to be applied to
ranked input lists, or to allow queryWorld to restrict Query ID
set}

\item{\code{mapUse}}{Boolean matrix indicating allowed mappings from the
Query ID namespace to the Ontology ID namespace. Computed after
applying minMapMatch to the raw Mapping matrix, and eliminating
rows (query input) and columns (ontology output) that lack
entries. Recursive elimination is also currently applied to
remove target IDs that have 'too small' fractional
represenation - this behavior may be removed.}

\item{\code{ontoUse}}{Boolean matrix indicating allowed associations
between IDs (either input IDs, or mapped IDs if a Mapping
matrix is used) and ontology terms. The matrix is filtered
using minOntoMatch (filters original raw matrix based on matrix
score), minOntoSize (minimum number of assigned terms required
to keep an ID), minSetSize (minimum number of IDs required to
keep a term) and maxSetSize (maximum number of IDs allowed for
a term to be kept)}

\item{\code{resultRaw}}{3D array holding raw p-values from phyper()
calculations, as well as i and n values for each calculation.}

\item{\code{resultAdj}}{p.adjust() processed values from resultRaw}

\item{\code{lastTop}}{data.frame of the most recent values from
topResults()}

\item{\code{isFiltered}}{Boolean flag tracking if filtering has already
been performed on the matrices}

\item{\code{logThresh}}{log10 value of the user's p-value significance
threshold}

\item{\code{roundUp}}{Integer representing the denominator of fractional
counts that should be rounded up to a full count. That is, a
value of 4 indicates that any fractional count >= 1/4 should
ultimately be considered a full count.}

\item{\code{pseudoRound}}{This is a pseudo-count based off of roundUp. It
is added to values before they are round()ed.}

\item{\code{filterLog}}{data.frame storing filtering events that transpire
during the pruning of matrices prior to analysis.}

\item{\code{ontoNames}}{Vector of all valid ontology terms following
filtering}

\item{\code{ontoSize}}{Total number of Ontology terms surviving filtering
= length(ontoNames).}

\item{\code{okInput}}{Standardized Query names, following matrix pruning}

\item{\code{idCount}}{Fractional totals for each ID used in the Ontology
matrix, after pruning}

\item{\code{ontoCount}}{Fractional total ID counts for each Ontology term
following pruning}

\item{\code{worldSize}}{Integer count of total number of IDs in the world,
following pruning}

\item{\code{bogusList}}{Used for debuging - holds the most recent list to
fail completely when run through processList()}

\item{\code{workSpace}}{Internal temporary structure used by topResult()}
}}

