% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetFisherLogger.R
\docType{class}
\name{SetFisherLogger-class}
\alias{SetFisherLogger}
\alias{SetFisherLogger-class}
\title{SetFisher Logger}
\description{
Utility class for messaging and recording of events
}
\section{Fields}{

\describe{
\item{\code{log}}{The data.table holding log messages}

\item{\code{useColor}}{Logical flag to indicate if color should be used in messaging}

\item{\code{verbose}}{Logical flag indicating if verbose messaging should be active}
}}

