% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetFisherMatrix.R
\docType{class}
\name{SetFisherMatrix-class}
\alias{SetFisherMatrix}
\alias{SetFisherMatrix-class}
\title{SetFisher Matrix}
\description{
Annotated sparse matrix for capturing query lists, identifier
mappings and ontology lookups
}
\section{Fields}{

\describe{
\item{\code{setfisher}}{Pointer back to the parent SetFisher object}

\item{\code{file}}{Path to file the matrix was loaded from}

\item{\code{fromRDS}}{Logical, true if the loaded file was an RDS object}

\item{\code{log}}{The SetFisherLogger object holding log (activity)
entries}

\item{\code{matrixRaw}}{The Matrix object as loaded from the file}

\item{\code{matrixUse}}{The Matrix object after any filters are applied}

\item{\code{matrixMD}}{data.table holding metadata associated with the
matrix}

\item{\code{levels}}{Character array of level names for factor matrices}

\item{\code{rowChanges}}{Named character vector of any row names that
needed changing. Values are the original name, names are the
names after processing with make.names() (if valid = TRUE) or
make.unique() if (valid = FALSE)}

\item{\code{colChanges}}{As per rowChanges, but for column names}
}}

