% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetFisherParamI.R
\docType{class}
\name{SetFisherParamI-class}
\alias{SetFisherParamI}
\alias{SetFisherParamI-class}
\title{SetFisher Paramter Interface}
\description{
Inherited class holding parameter manipulation functions
}
\section{Fields}{

\describe{
\item{\code{paramSet}}{List holding key-value pairs used by an object}

\item{\code{paramDef}}{List holding default key-value pairs}

\item{\code{varName}}{Single string holding extracted variable names. Used
for reporting R code that utilizes actual variable names being
used.}
}}

