% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{setFisherText}
\alias{setFisherText}
\title{Fancy Object Text}
\arguments{
\item{color}{Default \code{TRUE}, which will include ANSI color
tokens in the text for prettier (or messier, depending on your
perspective) text in terminals.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A character string.
}
\description{
Summarize the object in fancy, human-friendly way
}
\details{
\preformatted{
## Method Usage:
myObject$setFisherText( help=TRUE )

dt <- myObject$setFisherText(color=TRUE)
}

Create pretty-printed text, optionally colorized, to be used by the
RefClass \code{show()} method when summarizing the object.
}
