% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{topResults}
\alias{topResults}
\title{Top Results Report}
\arguments{
\item{\dots}{Passed to SetFisherAnalysis'
\link[SetFisherAnalysis]{topResults} method, as well as to
\link{analysisNames}.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
A data.frame aggregating the results from all
    analyses. This value will also be stored in the \link{lastTop}
    field.
}
\description{
Generate a data frame summarizing the top results from all analyses
}
\details{
\preformatted{
## Method Usage:
myObject$topResults( help=TRUE )

anaNames <- myObject$topResults( ... )
}

This method will consider all analysis objects held, and will
consolidate their results into a single data.frame. See
\link[SetFisherAnalysis]{topResults} (SetFisherAnalysis object) for
the recognized parameters that can influence the return value.
}
\seealso{
\link{lastTop}
}
