% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{tsvtable}
\alias{tsvtable}
\title{TSV Table}
\arguments{
\item{res}{Default \code{NA}, which will use the results in
\link{lastTop}, if available, or call \link{processAll} if
not. Otherwise the user may provide a specific data.frame to
process.}

\item{file}{Default \code{NA}, which will generate a temporary
file. Otherwise, a specified file path may be provided.}

\item{force}{Default \code{FALSE}. Passed on to \link{processAll}.}

\item{\dots}{Passed to \link{processAll} if \code{res} was not
provided.}

\item{help}{Default FALSE. If TRUE, show this help and perform no
other actions.}
}
\value{
The path to the TSV file generated.
}
\description{
Create a tab-separated value file from a top results data.frame
}
\details{
\preformatted{
## Method Usage:
myObject$tsvtable( help=TRUE )

tsvPath <- myObject$tsvtable( res=NA, file=NA, force=FALSE, ... )
}

Create a simple TSV file from a past analysis.
}
\seealso{
\link{topResults}
}
