% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsDocumentation.R
\name{weightMatrix}
\alias{weightMatrix}
\title{Generate Weight Matrix}
\description{
Generate the map mapWeights matrix if needed, and then return it
}
\details{
This method will return the \link{mapWeights} matrix if needed, and
then return it.

TODO - Extend documentation!!!
}
