% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetFisherAnalysis.R
\docType{class}
\name{SetFisherAnalysis-class}
\alias{SetFisherAnalysis-class}
\alias{SetFisherAnalysis}
\title{SetFisher Analysis}
\description{
Class object associating the components of an enrichment analysis
(query lists, ontology, and optional mapping matrix) with the
results
}
\section{Fields}{

\describe{
\item{\code{setfisher}}{Pointer back to the parent SetFisher object}

\item{\code{queryObj}}{AnnotatedMatrix representing the query list(s)}

\item{\code{queryWorld}}{Character verctor describing all objects in the
query namespace. Will default to all unique rows in the mapping
matrix, or all unique rows in the query matrix if no mapping
matrix is present.}

\item{\code{activeQueries}}{Character vector of normalized IDs
representing the world. Will either be taken from
\code{queryWorld}, or dynamically calculated from filtered /
intersected matrices.}

\item{\code{ontoObj}}{AnnotatedMatrix representing the mapping of IDs
(from the mapping matrix if present, otherwise from the query
matrix) to ontology terms}

\item{\code{mapObj}}{Optional AnnotatedMatrix mapping IDs in the Query
matrix to those used by the Ontology matrix}

\item{\code{mapWeights}}{Weight matrix used to accomodate multiple voting
from the Query namespace to the Ontology namespace. Taken from
the filtered (non-raw) Mapping Matrix if defined, or as a
simple identity matrix from the Query Matrix if not.}

\item{\code{ontoUse}}{Boolean matrix indicating allowed associations
between IDs (either input IDs, or mapped IDs if a Mapping
matrix is used) and ontology terms. The matrix is filtered to
align with mapWeights.}

\item{\code{queryOnto}}{A direct Query-to-Ontology matrix that takes into
account the existence or absence of a Mapping Matrix.}

\item{\code{resultRaw}}{3D array holding raw p-values from phyper()
calculations, as well as i and n values for each calculation.}

\item{\code{resultAdj}}{p.adjust() processed values from resultRaw}

\item{\code{lastTop}}{data.frame of the most recent values from
topResults()}

\item{\code{logThresh}}{log10 value of the user's p-value significance
threshold}

\item{\code{roundUp}}{Integer representing the denominator of fractional
counts that should be rounded up to a full count. That is, a
value of 4 indicates that any fractional count >= 1/4 should
ultimately be considered a full count.}

\item{\code{pseudoRound}}{This is a pseudo-count based off of roundUp. It
is added to values before they are round()ed.}

\item{\code{ontoNames}}{Vector of all valid ontology terms following
filtering}

\item{\code{ontoSize}}{Total number of Ontology terms surviving filtering
= length(ontoNames).}

\item{\code{okInput}}{Standardized Query names, following matrix pruning}

\item{\code{idCount}}{Fractional totals for each ID used in the Ontology
matrix, after pruning}

\item{\code{ontoCount}}{Fractional total ID counts for each Ontology term
following pruning}

\item{\code{worldSize}}{Integer count of total number of IDs in the world,
following pruning}

\item{\code{bogusList}}{Used for debuging - holds the most recent list to
fail completely when run through processList()}

\item{\code{workSpace}}{Internal temporary structure used by topResult()}
}}

\section{Methods}{

\describe{
\item{\code{initialize(params = NA, paramDefinitions = NA, help = FALSE, ...)}}{Create a new ParamSetI object; Invoked with ParamSetI(...)}

\item{\code{queryToOntology()}}{A weight matrix directly from query IDs to ontology terms}

\item{\code{show(...)}}{A wrapper for showLog}

\item{\code{weightMatrix()}}{Mutliple-voting weight matrix for mapping matrices, otherwise identity matrix}
}}

